7.2.1 Refactoring
